Use yaml for storing intercept responses
