+++ cmd/hetty/main.go
@@ -1,13 +1,9 @@
 package main
 
 import (
-	"bytes"
-	"compress/gzip"
 	"context"
 	"crypto/tls"
 	"flag"
-	"fmt"
-	"io/ioutil"
 	"log"
 	"net"
 	"net/http"
@@ -35,43 +31,6 @@ var (
 	addr       string
 	adminPath  string
 )
-
-func changeBody(res *http.Response, modifer func(body []byte) []byte) error {
-	body, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		return fmt.Errorf("Could not read response body: %v", err)
-	}
-
-	contentEncoding := res.Header.Get("Content-Encoding")
-
-	if contentEncoding == "" {
-		newBody := modifer(body)
-		res.Body = ioutil.NopCloser(bytes.NewBuffer(newBody))
-	}
-
-	if contentEncoding == "gzip" {
-		// TMP!
-		//res.Body = ioutil.NopCloser(bytes.NewBuffer(body))
-
-		gzipReader, err := gzip.NewReader(bytes.NewBuffer(body))
-		if err != nil {
-			return fmt.Errorf("Could not create gzip reader: %v", err)
-		}
-		defer gzipReader.Close()
-		body, err = ioutil.ReadAll(gzipReader)
-
-		// TODO: Gzip this body
-		newBody := modifer(body)
-		res.Header.Set("Content-Encoding", "")
-		res.Body = ioutil.NopCloser(bytes.NewBuffer(newBody))
-
-		if err != nil {
-			return fmt.Errorf("Could not read gzipped response body: %v", err)
-		}
-	}
-
-	return nil
-}
 
 func main() {
 	flag.StringVar(&caCertFile, "cert", "~/.hetty/hetty_cert.pem", "CA certificate filepath. Creates a new CA certificate is file doesn't exist")
@@ -134,34 +93,13 @@ func main() {
 	p.UseRequestModifier(reqLogService.RequestModifier)
 	p.UseResponseModifier(reqLogService.ResponseModifier)
 
-	p.UseResponseModifier(func(next proxy.ResponseModifyFunc) proxy.ResponseModifyFunc {
-		return func(res *http.Response) error {
-			if err := next(res); err != nil {
-				return err
-			}
+	err = proxy.NewInterceptor()
+	if err != nil {
+		log.Fatalf("[FATAL] Could not create Interceptor: %v", err)
+	}
 
-			if res.Request.URL.String() == "http://www.ue.wroc.pl/" {
-				//res.Header["X-Proxy"] = []string{"Hello"}
-
-				fmt.Println("Intercepting ue.wroc.pl request!")
-
-				err := changeBody(res, func(b []byte) []byte {
-					//fmt.Println(string(b))
-
-					prefix := []byte("<p>Pozdro poćwicz</p>")
-
-					return append(prefix, b...)
-				})
-				if err != nil {
-					panic(err)
-				}
-
-				//res.Body = ioutil.NopCloser(bytes.NewBuffer(body))
-			}
-
-			return nil
-		}
-	})
+	// TODO: Consider moving to proxy.NewProxy
+	p.UseResponseModifier(proxy.ResponseInterceptor)
 
 	var adminHandler http.Handler
 	if adminPath == "" {
